Split admin-firestore/remote.ts into one-concept-per-file
remote.ts (982 lines) aggregated the whole remote-sandbox arm: wire
type definitions, error translation, the write/read value codec,
channel dispatch, one-shot snapshots, onSnapshot listener plumbing,
Query state and construction, doc/collection refs, WriteBatch,
Transaction, and the createRemoteFirestore facade. Split into
sandbox/admin-firestore/remote/{wire-types,errors,value-codec,channel,
snapshots,listeners,query,doc-ref,batch,transaction,remote-firestore}.ts,
matching the one-concept-per-file shape of sandbox/firestore/admin-compat/.

Pure mechanical move: only relative import paths changed (one level
deeper) and cross-file exports added. No behavior change. The package
export (pyric/sandbox/admin-firestore) and its barrel are untouched in
this commit.
